Consolidate main application views in h.views package
This commit converts `h.views` from a module into a package, splitting
the various views into a set of modules:

- h.views.client
- h.views.exceptions
- h.views.help
- h.views.home
- h.views.main
